* new old filehandler for windows (see forum post) * fixed ambigious function calls in various files
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@6049 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Game/UI/GuiHandler.cpp
@@ -3807,7 +3807,7 @@ void CGuiHandler::DrawMiniMapMarker()
 	const float groundLevel = ground->GetHeight(camera->pos.x, camera->pos.z);
 
 	static float spinTime = 0.0f;
-	spinTime = fmodf(spinTime + gu->lastFrameTime, 60.0f);
+	spinTime = fmod(spinTime + gu->lastFrameTime, 60.0f);
 
 	glPushMatrix();
 	glTranslatef(camera->pos.x, groundLevel, camera->pos.z);
